requirements.txt
certifi==2018.11.29
Click==7.0
Django==2.1.5
django-crispy-forms==1.7.2
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.16.3
Pillow==6.0.0
pytz==2018.9
